==> 00_Data Preparation #5-RBI data.ipynb <==
import geopandas as gpd

data_rbi=gpd.read_file('data/input/RBI/jalan_RBI.gpkg')

data_prov=gpd.read_file('data/input/Master_Data/Prov_20201_ByIndo(1).shp')

data_rbi=data_rbi.sjoin(data_prov[['kdprov','geometry']],how='inner',op='intersects')

for i in data_rbi.kdprov:
    data_rbi[['kdprov','geometry']].query('kdprov==@i').to_file('data/input/RBI/'+i+'_allseasonedroad_rbi.gpkg',driver='GPKG')
