==> 00-01 Automating Script.ipynb <==
import papermill as pm

#from tqdm.contrib.concurrent import process_map

parameters = [{
    "kdprov": "11"},
    {"kdprov":"12"},
    {"kdprov":"13"},
    {"kdprov":"14"},
    {"kdprov":"15"},
    {"kdprov":"16"},
    {"kdprov":"17"},
    {"kdprov":"18"},
    {"kdprov":"19"},
    {"kdprov":"21"},
    {"kdprov":"31"},
    {"kdprov":"32"},
    {"kdprov":"33"},
    {"kdprov":"34"},
    {"kdprov":"35"},
    {"kdprov":"36"},
    {"kdprov":"51"},
    {"kdprov":"52"},
    {"kdprov":"53"},]

def run_exec(params):
    pm.execute_notebook(
        "00_Data Preparation #4 Rasterize Village Map.ipynb",
        f"00_Data Preparation #4-{params['kdprov']}.ipynb",
        parameters=params
    )

for params in parameters:
    run_exec(params)

parameters = [{
    "kdprov": "61"},
    {"kdprov":"62"},
    {"kdprov":"63"},
    {"kdprov":"64"},
    {"kdprov":"65"},
    {"kdprov":"71"},
    {"kdprov":"72"},
    {"kdprov":"73"},
    {"kdprov":"74"},
    {"kdprov":"75"},
    {"kdprov":"76"},
    {"kdprov":"81"},
    {"kdprov":"82"},
    {"kdprov":"91"},
    {"kdprov":"94"}]
for params in parameters:
    run_exec(params)

import papermill as pm
parameters = [
    #{"kdprov": "11"},
    #{"kdprov":"12"},
    #{"kdprov":"13"},
    #{"kdprov":"14"},
    #{"kdprov":"15"},
    #{"kdprov":"16"},
    #{"kdprov":"17"},
    #{"kdprov":"18"},
    #{"kdprov":"19"},
    #{"kdprov":"21"},
    #{"kdprov":"31"},
    #{"kdprov":"32"},
    #{"kdprov":"33"},
    #{"kdprov":"34"},
    #{"kdprov":"35"},
    #{"kdprov":"36"},
    #{"kdprov":"51"},
    #{"kdprov":"52"},
    #{"kdprov":"53"},
    #{"kdprov":"61"},
    #{"kdprov":"62"},
    #{"kdprov":"63"},
    #{"kdprov":"64"},
    #{"kdprov":"65"},
    #{"kdprov":"71"},
    #{"kdprov":"72"},
    #{"kdprov":"73"},
    #{"kdprov":"74"},
    #{"kdprov":"75"},
    #{"kdprov":"76"},
    #{"kdprov":"81"},
    #{"kdprov":"82"},
    #{"kdprov":"91"},
    {"kdprov":"94"}
    ]
def run_exec(params):
    pm.execute_notebook(
        "01_Population_Density #1 Preparing for 100m By KABKOTA.ipynb",
        f"01_Population Density #1-{params['kdprov']}.ipynb",
        parameters=params
    )
    pm.execute_notebook(
        "01_Population_Density #2 Resampling_to1km and Calculate BU Percentage-By Kabkota Rev.ipynb",
        f"01_Population Density #2-{params['kdprov']}.ipynb",
        parameters=params
    )
for params in parameters:
    run_exec(params)
