==> 02_DEGURBA #0 Persiapan.ipynb <==

import glob
from google.cloud import storage
from os.path import exists

#Download PopDens_100m
from google.cloud import storage

client = storage.Client()
ls_=list(client.list_blobs('bps-gcp-bucket', prefix='Degree of Urbanisation/PopDens'))
storage_client = storage.Client()
for i in ls_:
    if len(str(i).split('/'))==3:
        data_=str(i).split('/')[2].split(',')[0]
        print(data_)
        if (data_!='')and(exists('data/output/1_population_density/'+data_)==False) :
            bucket = storage_client.get_bucket('bps-gcp-bucket')
            source_blob_name='Degree of Urbanisation/PopDens/'+data_
            blob = bucket.blob(source_blob_name)
            blob.download_to_filename('data/output/1_population_density/'+data_)

#Download Fin_PopDens_1km
from google.cloud import storage

client = storage.Client()
ls_=list(client.list_blobs('bps-gcp-bucket', prefix='Degree of Urbanisation/Fin_1km_and_pulau_podens/'))
storage_client = storage.Client()
for i in ls_:
    if len(str(i).split('/'))==3:
        data_=str(i).split('/')[2].split(',')[0]
        print(data_)
        if (data_!='')and(exists('data/output/1_5_population_density/'+data_)==False) :
            bucket = storage_client.get_bucket('bps-gcp-bucket')
            source_blob_name='Degree of Urbanisation/Fin_1km_and_pulau_podens/'+data_
            blob = bucket.blob(source_blob_name)
            blob.download_to_filename('data/output/1_5_population_density/'+data_)

#Download OSM
from google.cloud import storage

client = storage.Client()
ls_=list(client.list_blobs('bps-gcp-bucket', prefix='Degree of Urbanisation/OSM/all-seasoned-road'))
storage_client = storage.Client()
for i in ls_:
    if len(str(i).split('/'))==4:
        data_=str(i).split('/')[3].split(',')[0]
        print(data_)
        if (data_!='')and(exists('data/input/OSM/'+data_)==False) :
            bucket = storage_client.get_bucket('bps-gcp-bucket')
            source_blob_name='Degree of Urbanisation/OSM/all-seasoned-road/'+data_
            blob = bucket.blob(source_blob_name)
            blob.download_to_filename('data/input/OSM/'+data_)

#Download raster data desa
from google.cloud import storage

client = storage.Client()
ls_=list(client.list_blobs('bps-gcp-bucket', prefix='Degree of Urbanisation/RasterDesa/'))
storage_client = storage.Client()
for i in ls_:
    if len(str(i).split('/'))==3:
        data_=str(i).split('/')[2].split(',')[0]
        print(data_)
        if (data_!='')and(exists('data/output/0_raster_data_desa/'+data_)==False) :
            bucket = storage_client.get_bucket('bps-gcp-bucket')
            source_blob_name='Degree of Urbanisation/RasterDesa/'+data_
            blob = bucket.blob(source_blob_name)
            blob.download_to_filename('data/output/0_raster_data_desa/'+data_)
