==> 02_DEGURBA #1  Deliniating Urban Centre, Urban Cluster, and Rural.ipynb <==
import geopandas as gpd

nama_file="00_fin_pulau_Bali_finalized.gpkg"

data_gpd=gpd.read_file("data/output/1_5_population_density/"+nama_file
                      ).reset_index()

data_gpd['Name']=data_gpd.index.astype(str).str.zfill(12)

##Check cluster centre
data_cl_centre=data_gpd.query('pop>=1500')

from pysal.lib import weights

wr = weights.contiguity.Rook.from_dataframe(data_cl_centre)
data_cl_centre['cluster_label']=wr.component_labels

import pandas as pd

tab_pop_u_centre_cand=pd.pivot_table(data_cl_centre,
                                     index='cluster_label',values='pop',aggfunc='sum')

list_u_centre=tab_pop_u_centre_cand.query('pop>50000').reset_index().cluster_label.tolist()

u_centre=data_cl_centre.query('cluster_label in @list_u_centre')

data_gpd['is_U_CENTRE']=data_gpd.Name.apply(lambda y: 1 if y in u_centre.Name.tolist() else 0)

#Filling gap of urbancentre => check 8 contiguity (queen)
init_u_centre=data_gpd.is_U_CENTRE.sum()
u_centre_up=0
print('N GRID WITH URBAN CENTRE:',init_u_centre)
while u_centre_up!=init_u_centre:
    init_u_centre=data_gpd.is_U_CENTRE.sum()
    wr = weights.contiguity.Queen.from_dataframe(data_gpd)
    data_gpd['neighbors']=wr.neighbors.values()
    temp=data_gpd.copy()
    for index, row in data_gpd.iterrows():
        if row['is_U_CENTRE']!=1:
            y_urbancentre=0
            if len(row['neighbors'])>4:
                for y in row['neighbors']:
                    if temp.loc[y,'is_U_CENTRE']==1:
                        y_urbancentre+=1
                if y_urbancentre>4:
                    data_gpd.at[index,'is_U_CENTRE']=1
    u_centre_up=data_gpd.is_U_CENTRE.sum()
    print('=====================')
    print('N GRID AFTER GAP FILLING:', u_centre_up)
print('==================')
u_centre=data_gpd.is_U_CENTRE.sum()
print('N GRID COMPLETED AFTER GAP FILLING',u_centre)

data_gpd['is_U_CENTRE_w_BU']=data_gpd.apply(lambda y:1 if (y.p_builtup>0.5)or(y.is_U_CENTRE==1) else 0,axis=1)

data_gpd=data_gpd[['Name','pop','geometry','p_builtup','is_U_CENTRE_w_BU']]

data_gpd

data_u_cluster=data_gpd.query('is_U_CENTRE_w_BU!=1').query('pop>=300')

wr = weights.contiguity.Queen.from_dataframe(data_u_cluster)
data_u_cluster['cluster_label']=wr.component_labels

tab_pop_u_cluster_cand=pd.pivot_table(data_u_cluster,
                                     index='cluster_label',values='pop',aggfunc='sum')
list_u_cluster=tab_pop_u_cluster_cand.query('pop>=5000').reset_index().cluster_label.tolist()

u_cluster=data_u_cluster.query('cluster_label in @list_u_cluster')

data_gpd['is_U_CLUSTER']=data_gpd.Name.apply(lambda y: 1 if y in u_cluster.Name.tolist() else 0)

data_gpd['classification']=data_gpd.apply(lambda y: 'CENTRE' 
                                          if y.is_U_CENTRE_w_BU==1 else ('CLUSTER' 
                                                                         if y.is_U_CLUSTER==1 else 'RURAL'),axis=1)

data_gpd.classification.unique()

nmdegurba_=nama_file.replace("00_fin_pulau_","").replace("_finalized.gpkg","")

data_gpd.to_file('data/output/2_grid_degurba/degurba_'+nmdegurba_+'_00.gpkg',driver='GPKG')
