==> 02_DEGURBA #2 Deliniating Small Spatial Unit.ipynb <==
import geopandas as gpd
import pandas as pd
from glob import glob

kdprov=['51']
nmfiledegurba="degurba_Bali_00.gpkg"

data_degurba_grod=gpd.read_file('data/output/2_grid_degurba/'+nmfiledegurba)

data_pop=gpd.GeoDataFrame()
for i in kdprov:
    gl_pop=glob("data/output/1_population_density/grid_pop_"+i+"*_100m.gpkg")
    for j in gl_pop:
        y_temp=gpd.read_file(j)
        y_temp['geometry']=y_temp.centroid
        y_temp=y_temp[['j_pend','geometry','p_area']]
        data_pop=y_temp.append(data_pop)

data_desa_raster=gpd.GeoDataFrame()
for i in kdprov:
    temp=gpd.read_file('data/output/0_raster_data_desa/'+'desa_pend_rasterize_'+i+'.gpkg')
    data_desa_raster=temp.append(data_desa_raster)

pop_w_grid=data_pop.sjoin_nearest(data_degurba_grod)[['j_pend','classification','geometry']]

data_desa=data_desa_raster.sjoin(pop_w_grid)[['j_pend','classification','iddesa','geometry']]

data_desa

tab_data=pd.pivot_table(data_desa,index='iddesa',columns='classification',
                        values='j_pend',aggfunc='sum').fillna(0).reset_index()

l_col=['CENTRE','CLUSTER','RURAL']
for i in l_col:
    if i not in tab_data.columns:
        tab_data[i]=0
tab_data

def define_LAU(y):
    ['CENTRE','CLUSTER','RURAL']
    total_pop=y.CENTRE+y.CLUSTER+y.RURAL
    if total_pop==0:
        return 'RURALS'
    elif (y.CENTRE/total_pop)>=0.5:
        return 'CITIES'
    elif (((y.CENTRE/total_pop)<0.5)&((y.RURAL/total_pop)<0.5)):
        return 'TOWNS'
    else:
        return 'RURALS'

tab_data['LAU']=tab_data.apply(lambda y:define_LAU(y),axis=1)

peta_desa=gpd.read_file(r"data/input/Master_Data/DESA_20201_BYINDO.gpkg").query('kdprov in @kdprov')

peta_desa_LAU=peta_desa.merge(tab_data[['iddesa','LAU']],how='left').fillna('RURALS')

for i in kdprov:
    peta_desa_LAU.query("kdprov==@i").to_file(
        'data/output/3_small_spatial_unit_degurba/small_spatial_unit_'+i+'.gpkg',driver='GPKG')

from dbfread import DBF
mfd_=pd.DataFrame(iter(DBF('data/input/Master_Data/mfd_2020_1.dbf')))

mfd_['iddesa']=mfd_.PROP+mfd_.KAB+mfd_.KEC+mfd_.DESA

for i in kdprov:
    peta_desa_LAU.merge(mfd_[['iddesa','DRH']]).query("kdprov==@i").to_file(
        'data/output/3_small_spatial_unit_degurba/small_spatial_unit_'+i+'.gpkg',driver='GPKG')

peta_desa_LAU.merge(mfd_[['iddesa','DRH']]).pivot_table(index='DRH',columns='LAU',values='iddesa',aggfunc='count').fillna(0)
