==> 03_RAI #1 RAI OSM Calculation by kabkota.ipynb <==
import geopandas as gpd
import pandas as pd
import warnings
from tqdm.contrib.concurrent import process_map
warnings.filterwarnings('ignore')
from multiprocessing import Process, Manager

kdprov='51'

data_rai=gpd.read_file('data/input/OSM/'+kdprov+'_allseasonedroad_OSM.gpkg').to_crs('ESRI:54034').dissolve()

nmfiledegurba="degurba_Bali_00.gpkg"
gf=gpd.read_file("data/output/2_grid_degurba/"+nmfiledegurba).to_crs('ESRI:54034')
gf['RAI']=0

data_desa=gpd.read_file('data/input/Master_Data/DESA_20201_BYINDO.gpkg').to_crs('ESRI:54034').query('kdprov==@kdprov')
gf=gf.sjoin(data_desa[['geometry']])[gf.columns]

def data_masking_degurba(i):
    global degurba_desa,data_highway,gf,affected_RAI
    degurba_desa=gpd.sjoin(gf,data_desa.query('iddesa==@i'))
    data_highway_j=data_rai.overlay(degurba_desa,how='intersection')[['geometry']]
    data_highway_j['reg']=1
    data_highway_j=data_highway_j.dissolve(by='reg')
    data_highway_j.geometry=data_highway_j.buffer(2000)
    nm_rai=gf.sjoin(data_highway_j)['Name'].tolist()
    affected_RAI.extend(nm_rai)
    
    #print(affected_RAI)

for j in data_desa.idkab.unique():
    print('Running for ',j)
    data_desa_kab=data_desa.query('idkab==@j')
    manager=Manager()
    affected_RAI=manager.list()
    process_map(data_masking_degurba,data_desa_kab.iddesa.unique())
    gf[gf.Name.isin(list(affected_RAI))].to_file('data/output/4_RAI/rai_osm_degurba_'+j+'.gpkg',driver='GPKG')

from glob import glob
degurba_rai=glob('data/output/4_RAI/rai_osm_degurba_'+kdprov+'*.gpkg')

data_=gpd.GeoDataFrame()
l_name=[]
for i in degurba_rai:
    temp=gpd.read_file(i)['Name']
    l_name.extend(list(temp))
l_name=list(set(l_name))

gf['accessed']=gf.Name.apply(lambda y: 1 if y in l_name else 0)

gf.to_file('data/output/4_RAI/fin_rai_osm_degurba_'+kdprov+'.gpkg')

data_p=gf.pivot_table(index='classification',columns='accessed',values='pop',aggfunc='sum').fillna(0)

data_p['percent']=data_p[1]*100/(data_p[0]+data_p[1])

data_p
