==> 04_RAI_DATA-Copy2.ipynb <==
from pyrosm import OSM
from pyrosm import get_data

import geopandas as gpd
import pandas as pd

jalan=gpd.read_file('jalan_nusra.gpkg')

jalan.head()

highway_l=['secondary', 'tertiary', 'primary',
       'trunk','motorway' ]
surface_l=['paved', 'asphalt', 'unpaved', None,'cobblestone', 
       'paving_stones', 'gravel;asphalt', 'concrete', 'asphalt_ruts',
       'Hotmix', 'dirt;asphalt', 'metal',
       'sett', 'concrete:plates',
       'compacted', 'Aspal dan Tanah',
       'permukaan tanah dan Beton', 'Tanah dan Aspal', 'Aspal dan Beton', 'Tanah dan Beton',
       'concrete:lanes', 'Rabat', 'Hotmix dan Tanah',
       'paving_stones:30', 'driveway', 'unhewn_cobblestone']
jalan.query('highway in @highway_l').query('surface in @surface_l').to_file('data_jalan_nusra.gpkg',driver='GPKG')

# Initialize the OSM parser object
osm = OSM("nusa-tenggara-latest-internal.osm.pbf")

drive_net = osm.get_network(network_type="driving")

drive_net.surface.unique()

drive_net.highway.unique()

drive_net.to_file('data_jalan_osm.gpkg',driver='GPKG')

gf=gpd.read_file('degurba_5300_rev_w_buitup_adjustment_3.gpkg')

highway_=drive_net.query('highway in @highway_l').query('surface in @surface_l')

data_highway=highway_.overlay(gf.dissolve(),how='intersection')

high_2km=data_highway.to_crs(3857).buffer(2000).to_crs(4326)

data_highway['geometry']=high_2km

data_jalan=data_highway[['geometry']]
data_jalan['reg']=1

data_jalan.to_file('jalan_2km_53.gpkg',driver='GPKG')

data_int=gpd.sjoin(gf,data_jalan,how='left',predicate='intersects')

data_int.reg.unique()

unique_nama=data_int.query('reg==1')['Name'].unique()

gf['accessed']=gf.Name.apply(lambda y:1 if y in unique_nama else 0)

data_p=gf.pivot_table(index='classification',columns='accessed',values='pop_built_s2',aggfunc='sum').fillna(0)

data_p['percent']=data_p[1]*100/(data_p[0]+data_p[1])

data_p

gf.to_file('rai_degurba_ntt.gpkg',driver='GPKG')
